코드 가독성과 유지 보수를 위한 GetHandler 작성
