+++ src/main/java/handler/PostHandler.java
@@ -1,0 +1,14 @@
+package handler;
+
+public class PostHandler
+{
+    private PostHandler() {}
+
+    private static class LazyHolder{
+        private static final PostHandler INSTANCE = new PostHandler();
+    }
+    public static PostHandler getInstance()
+    {
+        return LazyHolder.INSTANCE;
+    }
+}
