+++ src/main/java/util/FileDetection.java
@@ -23,8 +23,6 @@ public class FileDetection
             e.printStackTrace();
         }
     }
-
-
     public static String getPath(String path)
     {
         File file = new File(path);
